fix: newsfeed wrappers
