+++ src/components/NewsFeed/index.tsx
@@ -1,18 +1,23 @@
 import React, { useContext } from 'react'
 import { NewsFeed } from '@gooddollar/good-design'
-import { NewsFeedContext } from '@gooddollar/web3sdk-v2'
+import { NewsFeedContext, NewsFeedProvider } from '@gooddollar/web3sdk-v2'
 
-export const feedConfig = {
-    production: {
-        feedFilter: {
-            context: process.env.REACT_APP_FEEDCONTEXT_PROD,
-            tag: 'publishDapp',
-        },
-    },
-}
+import { feedConfig } from 'constants/config'
+import { getNetworkEnv } from 'utils/env'
 
 export const NewsFeedWidget = ({ variant, direction = 'row' }: { variant?: string; direction?: 'row' | 'column' }) => {
     const { feed } = useContext(NewsFeedContext)
 
     return <NewsFeed {...{ feed, variant, direction }} />
+}
+
+export const NewsFeedWrapper = ({ children }) => {
+    const network = getNetworkEnv()
+    const prodOrQa = /\b(production|staging)\b/.test(network)
+
+    return (
+        <NewsFeedProvider {...(prodOrQa ? { feedFilter: feedConfig.production.feedFilter } : { env: 'qa' })}>
+            {children}
+        </NewsFeedProvider>
+    )
 }
