+++ src/constants/config.ts
@@ -1,0 +1,8 @@
+export const feedConfig = {
+    production: {
+        feedFilter: {
+            context: process.env.REACT_APP_FEEDCONTEXT_PROD,
+            tag: 'publishDapp',
+        },
+    },
+}
